requirements.txt
fastapi
uvicorn
pydantic
groq
python-dotenv
uvicorn
python-jose
passlib